feat(mobile): fullscreen surfaces and full-width sessions drawer
